historian bug fix

+++ YT-RED/Logging/Historian.cs
@@ -12,13 +12,9 @@ namespace YT_RED.Logging
     public static class Historian
     {
         private static string historyFile = "history.json";
+        public static bool Loaded { get { return historyWasLoaded; } }
         private static bool historyWasLoaded = false;
-        public static List<DownloadLog> DownloadHistory;
-
-        public static void Init()
-        {
-            DownloadHistory = new List<DownloadLog>();
-        }
+        public static List<DownloadLog> DownloadHistory = new List<DownloadLog>();
 
         public static async Task<bool> LoadDownloadHistory()
         {
@@ -108,7 +104,7 @@ namespace YT_RED.Logging
                 {
                     DownloadHistory.RemoveAll(h => h.Downloaded.Date < DateTime.Today.AddDays(-AppSettings.Default.General.HistoryAge).Date);
                 }
-                await SaveHistory();
+                bool cleaned = await SaveHistory();
             }
             catch(Exception ex)
             {
@@ -143,6 +139,11 @@ namespace YT_RED.Logging
             try
             {
                 var json = JsonConvert.SerializeObject(DownloadHistory, Formatting.Indented);
+                FileInfo hf = new FileInfo(historyFile);
+                if(await IsFileLocked(hf))
+                {
+                    return false;
+                }
                 await Task.Run(() => File.WriteAllText(historyFile, json));
                 return true;
             }
@@ -151,6 +152,31 @@ namespace YT_RED.Logging
                 ExceptionHandler.LogException(ex);
             }
             return false;
+        }
+
+        private static async Task<bool> IsFileLocked(FileInfo file)
+        {
+            return await Task.Run(() =>
+            {
+                try
+                {
+                    using (FileStream stream = file.Open(FileMode.Open, FileAccess.Read, FileShare.None))
+                    {
+                        stream.Close();
+                    }
+                }
+                catch (IOException)
+                {
+                    //the file is unavailable because it is:
+                    //still being written to
+                    //or being processed by another thread
+                    //or does not exist (has already been processed)
+                    return true;
+                }
+
+                //file is not locked
+                return false;
+            });
         }
     }
